graph as a template + structure for graph files

+++ cpp/oneapi/dal/data/undirected_adjacency_array_graph.hpp
@@ -19,11 +19,9 @@
 #include <cstdint>
 #include <memory>
 
-#include "oneapi/dal/common.hpp"
-#include "oneapi/dal/data/array.hpp"
 #include "oneapi/dal/data/detail/graph_container.hpp"
 #include "oneapi/dal/data/detail/undirected_adjacency_array_graph_impl.hpp"
-#include "oneapi/dal/detail/common.hpp"
+#include "oneapi/dal/data/graph_common.hpp"
 
 namespace oneapi::dal::preview {
 
@@ -43,7 +41,7 @@ const typename G::pimpl &get_impl(const G &g) {
 template <typename VertexValue = empty_value,
           typename EdgeValue   = empty_value,
           typename GraphValue  = empty_value,
-          typename IndexType   = std::int64_t,
+          typename IndexType   = std::int32_t,
           typename Allocator   = std::allocator<empty_value>>
 class undirected_adjacency_array_graph {
 public:
@@ -180,7 +178,7 @@ void convert_to_csr_impl(const edge_list<vertex_type<G>> &edges, G &g);
 template <typename VertexValue = empty_value,
           typename EdgeValue   = empty_value,
           typename GraphValue  = empty_value,
-          typename IndexType   = std::int64_t,
+          typename IndexType   = std::int32_t,
           typename Allocator   = std::allocator<empty_value>>
 using undirected_adjacency_array =
     undirected_adjacency_array_graph<VertexValue, EdgeValue, GraphValue, IndexType, Allocator>;
